fix table loki specs
