+++ frontend/test/__support__/storybook.tsx
@@ -1,4 +1,8 @@
 // Storybook helpers
+// @ts-expect-error There is no type definition
+import createAsyncCallback from "@loki/create-async-callback";
+import type { StoryFn } from "@storybook/react";
+import { useEffect, useMemo } from "react";
 
 import { SdkThemeProvider } from "embedding-sdk/components/private/SdkThemeProvider";
 import type { MetabaseTheme } from "metabase/embedding-sdk/theme";
@@ -93,3 +97,19 @@ export const IsomorphicVisualizationStory = ({
     </Box>
   );
 };
+
+export function createWaitForResizeToStopDecorator(timeoutMs: number = 1000) {
+  return function WaitForResizeToStopDecorator(Story: StoryFn) {
+    const asyncCallback = useMemo(() => createAsyncCallback(), []);
+
+    useEffect(() => {
+      const timeoutId = setTimeout(asyncCallback, timeoutMs);
+
+      return () => {
+        clearTimeout(timeoutId);
+      };
+    }, [asyncCallback]);
+
+    return <Story />;
+  };
+}
